tools/requirements.txt
RPi.GPIO
psycopg2-binary
python-decouple
argon2-cffi
mfrc522